Refactor map options drawer into separate part
